Move registry helpers into pkg/registry
