+++ pkg/registry/apps/deployment/strategy.go
@@ -29,9 +29,9 @@ import (
 	"k8s.io/apiserver/pkg/registry/rest"
 	"k8s.io/apiserver/pkg/storage/names"
 	"k8s.io/kubernetes/pkg/api/legacyscheme"
-	"k8s.io/kubernetes/pkg/api/pod"
 	"k8s.io/kubernetes/pkg/apis/extensions"
 	"k8s.io/kubernetes/pkg/apis/extensions/validation"
+	podhelpers "k8s.io/kubernetes/pkg/registry/core/pod/helpers"
 )
 
 // deploymentStrategy implements behavior for Deployments.
@@ -70,7 +70,7 @@ func (deploymentStrategy) PrepareForCreate(ctx genericapirequest.Context, obj ru
 	deployment.Status = extensions.DeploymentStatus{}
 	deployment.Generation = 1
 
-	pod.DropDisabledAlphaFields(&deployment.Spec.Template.Spec)
+	podhelpers.DropDisabledPodSpecAlphaFields(&deployment.Spec.Template.Spec)
 }
 
 // Validate validates a new deployment.
@@ -94,8 +94,8 @@ func (deploymentStrategy) PrepareForUpdate(ctx genericapirequest.Context, obj, o
 	oldDeployment := old.(*extensions.Deployment)
 	newDeployment.Status = oldDeployment.Status
 
-	pod.DropDisabledAlphaFields(&newDeployment.Spec.Template.Spec)
-	pod.DropDisabledAlphaFields(&oldDeployment.Spec.Template.Spec)
+	podhelpers.DropDisabledPodSpecAlphaFields(&newDeployment.Spec.Template.Spec)
+	podhelpers.DropDisabledPodSpecAlphaFields(&oldDeployment.Spec.Template.Spec)
 
 	// Spec updates bump the generation so that we can distinguish between
 	// scaling events and template changes, annotation updates bump the generation
