+++ pkg/registry/apps/statefulset/BUILD
@@ -16,9 +16,9 @@ go_library(
     importpath = "k8s.io/kubernetes/pkg/registry/apps/statefulset",
     deps = [
         "//pkg/api/legacyscheme:go_default_library",
-        "//pkg/api/pod:go_default_library",
         "//pkg/apis/apps:go_default_library",
         "//pkg/apis/apps/validation:go_default_library",
+        "//pkg/registry/core/pod/helpers:go_default_library",
         "//vendor/k8s.io/api/apps/v1beta1:go_default_library",
         "//vendor/k8s.io/api/apps/v1beta2:go_default_library",
         "//vendor/k8s.io/apimachinery/pkg/api/equality:go_default_library",
