+++ pkg/registry/core/pod/BUILD
@@ -15,11 +15,11 @@ go_library(
     importpath = "k8s.io/kubernetes/pkg/registry/core/pod",
     deps = [
         "//pkg/api/legacyscheme:go_default_library",
-        "//pkg/api/pod:go_default_library",
         "//pkg/apis/core:go_default_library",
         "//pkg/apis/core/helper/qos:go_default_library",
         "//pkg/apis/core/validation:go_default_library",
         "//pkg/kubelet/client:go_default_library",
+        "//pkg/registry/core/pod/helpers:go_default_library",
         "//vendor/k8s.io/apimachinery/pkg/api/errors:go_default_library",
         "//vendor/k8s.io/apimachinery/pkg/api/meta:go_default_library",
         "//vendor/k8s.io/apimachinery/pkg/apis/meta/v1:go_default_library",
@@ -71,6 +71,7 @@ filegroup(
     name = "all-srcs",
     srcs = [
         ":package-srcs",
+        "//pkg/registry/core/pod/helpers:all-srcs",
         "//pkg/registry/core/pod/rest:all-srcs",
         "//pkg/registry/core/pod/storage:all-srcs",
     ],
