+++ pkg/registry/settings/podpreset/BUILD
@@ -15,9 +15,9 @@ go_library(
     importpath = "k8s.io/kubernetes/pkg/registry/settings/podpreset",
     deps = [
         "//pkg/api/legacyscheme:go_default_library",
-        "//pkg/api/pod:go_default_library",
         "//pkg/apis/settings:go_default_library",
         "//pkg/apis/settings/validation:go_default_library",
+        "//pkg/registry/core/pod/helpers:go_default_library",
         "//vendor/k8s.io/apimachinery/pkg/apis/meta/internalversion:go_default_library",
         "//vendor/k8s.io/apimachinery/pkg/apis/meta/v1:go_default_library",
         "//vendor/k8s.io/apimachinery/pkg/runtime:go_default_library",
