+++ pkg/registry/settings/podpreset/strategy.go
@@ -22,9 +22,9 @@ import (
 	genericapirequest "k8s.io/apiserver/pkg/endpoints/request"
 	"k8s.io/apiserver/pkg/storage/names"
 	"k8s.io/kubernetes/pkg/api/legacyscheme"
-	"k8s.io/kubernetes/pkg/api/pod"
 	"k8s.io/kubernetes/pkg/apis/settings"
 	"k8s.io/kubernetes/pkg/apis/settings/validation"
+	podhelpers "k8s.io/kubernetes/pkg/registry/core/pod/helpers"
 )
 
 // podPresetStrategy implements verification logic for Pod Presets.
@@ -46,7 +46,7 @@ func (podPresetStrategy) PrepareForCreate(ctx genericapirequest.Context, obj run
 	pip := obj.(*settings.PodPreset)
 	pip.Generation = 1
 
-	pod.DropDisabledVolumeMountsAlphaFields(pip.Spec.VolumeMounts)
+	podhelpers.DropDisabledVolumeMountsAlphaFields(pip.Spec.VolumeMounts)
 }
 
 // PrepareForUpdate clears fields that are not allowed to be set by end users on update.
@@ -54,8 +54,8 @@ func (podPresetStrategy) PrepareForUpdate(ctx genericapirequest.Context, obj, ol
 	newPodPreset := obj.(*settings.PodPreset)
 	oldPodPreset := old.(*settings.PodPreset)
 
-	pod.DropDisabledVolumeMountsAlphaFields(oldPodPreset.Spec.VolumeMounts)
-	pod.DropDisabledVolumeMountsAlphaFields(newPodPreset.Spec.VolumeMounts)
+	podhelpers.DropDisabledVolumeMountsAlphaFields(oldPodPreset.Spec.VolumeMounts)
+	podhelpers.DropDisabledVolumeMountsAlphaFields(newPodPreset.Spec.VolumeMounts)
 
 	// Update is not allowed
 	newPodPreset.Spec = oldPodPreset.Spec
